requirements.txt
requests==2.31.0
python-dotenv==1.0.0
fastapi==0.103.1
uvicorn==0.23.2
APScheduler==3.10.4
openai==1.3.0
sqlalchemy==2.0.20
python-multipart==0.0.6
Pillow==10.0.0 